Added various ember features to start front end
Created an ember model for tasks that handles the data from
the front end and pushes it to the back end. There is currently form
at `/tasks` where you create a task. Currently, this is being
used with a `task-list` component but right now the data is not
being handled properly on the front end
The same information applies for bands, created a route and a component
for bands