Added more components to make code a bit more modular
Added a `band-list/card` component that as of right now is just
the band title, this will link to tasks.
Also added a component for a new band form.
Added a component to be able to delete tasks that will eventually
have an edit button as well so a user can edit tasks as well.