Added delete and update functionality for bands
The bands route now has a sub-route that goes to edit
I also refactored how bands display on the page. So the band
now reference the bands template.
The way update works is there is a `card` component that is linked
to the titles of the band. When you click on update, the action is
sent up to `bands/template/route.js` file. Then, update uses
`transitionTo` to go to the `band-list/modify` component which
has a form to accept the edit.