Added Contacts Route, each contact has CRUD
Added a contacts route that display all contacts along with
a form to create a new one. Each contact has a button that allows
the user to delete and update said contact. The update button works
by transitioning to a different route and upon successful update
re-routes back to contacts