Began setting up ownership of Band information
Set up the beginning stages of having bands own their tasks. This
will assist in setting up the same functionality for expenses and
contacts. Added a band/task route that is currently not working but
does have an open issue which will aid in finishing that setup. Added
a 'go-to' button that renders with each band, currently that will go
to that bands Dashboard which just results in a link to their tasks

Also set up error handling for when a band is not created properly
by using a promise chain. Also routed user to `/bands` on login