Get tasks to post successfully with their associated bands
Discovered that since, in the router, `/tasks` no longer exists,
actions aren't sent there anymore, they just go straight up to
`/band/tasks`. So, in the component, you need to set the band
associated with the task and then you send the action up
and it is saved through an adapter which was also created.