Address and Attempt to resolve Bugs
A caching bug came to light wherein when something is deleted,
it is cached by ember. An example is deleting a contact and then
adding another contact. Adding this contact prompts the previously
deleted contact to return. This is system wide. Attempted to resolve
this issue by transitioning to a different route in the promise chain
that handles create. Also tried this in the function that handles
delete. Also tried moving the create form onto a new route but the
issue persists here as well.

Added various styling while taking breaks from trying to resolve
issues